support add commit pool and height to block

+++ src/pools/tx_pool_manager.h
@@ -195,8 +195,8 @@ private:
     uint64_t prev_check_leader_valid_ms_ = 0;
     uint64_t prev_cacultate_leader_valid_ms_ = 0;
     std::shared_ptr<sync::KeyValueSync> kv_sync_ = nullptr;
-    uint32_t prev_synced_pool_index_ = 0;
-    uint32_t prev_sync_height_tree_tm_ms_ = 0;
+    uint64_t prev_synced_pool_index_ = 0;
+    uint64_t prev_sync_height_tree_tm_ms_ = 0;
     volatile uint64_t synced_max_heights_[common::kInvalidPoolIndex] = { 0 };
     volatile uint64_t cross_synced_max_heights_[network::kConsensusWaitingShardOffset] = { 0 };
     uint64_t latest_elect_height_ = 0;
